+++ test/test51.cxx
@@ -46,7 +46,7 @@ void test_051()
 	0,
 	"Bad position after seeking to beginning of large object.");
 
-      A.write(Buf, static_cast<int>(Contents.size()));
+      A.write(Buf, Contents.size());
       A.seek(0, std::ios::beg);
       PQXX_CHECK_EQUAL(
 	size_t(A.read(Buf, Size)),
